+++ src/main/java/com/braintreegateway/ThreeDSecureInfo.java
@@ -15,6 +15,11 @@ public class ThreeDSecureInfo {
     private String threeDSecureVersion;
     private String dsTransactionId;
     private String threeDSecureAuthenticationId;
+    private String acsTransactionId;
+    private String paresStatus;
+    private String threeDSecureServerTransactionId;
+    private ThreeDSecureLookupInfo threeDSecureLookupInfo;
+    private ThreeDSecureAuthenticateInfo threeDSecureAuthenticateInfo;
 
     public ThreeDSecureInfo(NodeWrapper node) {
         liabilityShifted = node.findBoolean("liability-shifted");
@@ -27,6 +32,18 @@ public class ThreeDSecureInfo {
         threeDSecureVersion = node.findString("three-d-secure-version");
         dsTransactionId = node.findString("ds-transaction-id");
         threeDSecureAuthenticationId = node.findString("three-d-secure-authentication-id");
+        acsTransactionId = node.findString("acs-transaction-id");
+        paresStatus = node.findString("pares-status");
+        threeDSecureServerTransactionId = node.findString("three-d-secure-server-transaction-id");
+
+        NodeWrapper threeDSecureLookupInfoNode = node.findFirst("lookup");
+        if (threeDSecureLookupInfoNode != null && !threeDSecureLookupInfoNode.isBlank()) {
+            threeDSecureLookupInfo = new ThreeDSecureLookupInfo(threeDSecureLookupInfoNode);
+        }
+        NodeWrapper threeDSecureAuthenticateInfoNode = node.findFirst("lookup");
+        if (threeDSecureAuthenticateInfoNode != null && !threeDSecureAuthenticateInfoNode.isBlank()) {
+            threeDSecureAuthenticateInfo = new ThreeDSecureAuthenticateInfo(threeDSecureAuthenticateInfoNode);
+        }
     }
 
     public ThreeDSecureInfo(Map<String, Object> map) {
@@ -40,6 +57,18 @@ public class ThreeDSecureInfo {
         threeDSecureVersion = (String) map.get("threeDSecureVersion");
         dsTransactionId = (String) map.get("dsTransactionId");
         threeDSecureAuthenticationId = (String) map.get("threeDSecureAuthenticationId");
+        acsTransactionId = (String) map.get("acsTransactionId");
+        paresStatus = (String) map.get("paresStatus");
+        threeDSecureServerTransactionId = (String) map.get("threeDSecureServerTransactionId");
+
+        Map<String, Object> threeDSecureLookupInfoMap = (Map) map.get("lookup");
+        if (threeDSecureLookupInfoMap != null) {
+            threeDSecureLookupInfo = new ThreeDSecureLookupInfo(threeDSecureLookupInfoMap);
+        }
+        Map<String, Object> threeDSecureAuthenticateInfoMap = (Map) map.get("lookup");
+        if (threeDSecureAuthenticateInfoMap != null) {
+            threeDSecureAuthenticateInfo = new ThreeDSecureAuthenticateInfo(threeDSecureAuthenticateInfoMap);
+        }
     }
 
     public String getStatus() {
@@ -80,5 +109,25 @@ public class ThreeDSecureInfo {
 
     public String getThreeDSecureAuthenticationId() {
         return threeDSecureAuthenticationId;
+    }
+
+    public String getAcsTransactionId() {
+        return acsTransactionId;
+    }
+
+    public String getParesStatus() {
+        return paresStatus;
+    }
+
+    public String getThreeDSecureServerTransactionId() {
+        return threeDSecureServerTransactionId;
+    }
+
+    public ThreeDSecureLookupInfo getThreeDSecureLookupInfo() {
+        return threeDSecureLookupInfo;
+    }
+
+    public ThreeDSecureAuthenticateInfo getThreeDSecureAuthenticateInfo() {
+        return threeDSecureAuthenticateInfo;
     }
 }